add_unique_index_to_friendships and edit views

+++ app/assets/javascripts/add_friend_to_list.js
@@ -7,16 +7,21 @@ function addFriendToList(){
         data: { friend_email:friend, authenticity_token:$('meta[name="csrf-token"]').attr("content")},
         dataType: "json",
         success: function (response) {
-          // console.log(response['img'].thumb.url)
-            $("#friendlist").append('<div class="col-lg-4 col-sm-6 portfolio-item">'
-                +'<img src='+response['img'].thumb.url+'/>'
-                +' '+response['name']+' '
-                +'<a data-confirm="Are you sure?" class="my-link" data-remote="true" rel="nofollow" data-method="delete" href="/friendships/'+response['id']+'">Unfriend</a>'
-                // +'<a href="" class="my-link" onclick="RemoveFriend()" >Unfriend</a>'+'<br /><br /><br /></div>'
+            if (response['found']){
+                alert("Already Friend")
+            }else {
+                // console.log(response['img'].thumb.url)
+                $("#friendlist").append('<div class="col-lg-4 col-sm-6 portfolio-item">'
+                    + '<img src=' + response['img'].thumb.url + '/>'
+                    + ' ' + response['name'] + ' '
+                    + '<a data-confirm="Are you sure?" class="my-link" data-remote="true" rel="nofollow" data-method="delete" href="/friendships/' + response['id'] + '">Unfriend</a>'
+                    // +'<a href="" class="my-link" onclick="RemoveFriend()" >Unfriend</a>'+'<br /><br /><br /></div>'
                 )
-             $(".my-link").bind('ajax:success',function(){$(this).parent().remove();  
-                console.log(response.msg)
-                   })
+                $(".my-link").bind('ajax:success', function () {
+                    $(this).parent().remove();
+                    console.log(response.msg)
+                })
+            }
         } 
     });
